Refactor: extract the code for adapting the docker mount path
